js/src/xpconnect/tests/chrome/test_bug448587.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=448587.xul
-->
<window title="Mozilla Bug 503926"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=448587"
     target="_blank">Mozilla Bug 448587</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 448587 **/
  const Cu = Components.utils;
  var sandbox = new Cu.Sandbox("about:blank");
  var fwrapper = Cu.evalInSandbox("function f() {} f", sandbox);
  ok(fwrapper.prototype == Cu.evalInSandbox("f.prototype", sandbox));
  ]]>
  </script>
</window>
